fix: cap for quests on front-end
front-end wrongly assumed `weeklyLimit` to be the max amount of points a user can get per week for the quest, whereas it should be interpreted as the maximum items the user can receive points for.

example:
`weeklyLimit` of 5 with 100 points/item, should give 500 max. points. Previously it gave 5 points / week max.